refactor: rename claude.ts to prompts.ts and generalize AI terminology
- Rename claude.ts/test to prompts.ts/test
- Update all imports and documentation references
- Generalize 'Claude' to 'AI' in user-facing messages and templates
- Fix version variable reference in src/index.ts
- Update package.json description
